change size_t for uint64 in order to avoid architecture change problems

+++ Engine/Component/ComponentMesh.h
@@ -33,7 +33,7 @@ private:
 
 public:
 	std::vector<Vertex> vertices;
-	std::vector<unsigned int> indices;
+	std::vector<UINT64> indices;
 	unsigned int material_index = -1;
 
 private:
